requirements.txt

streamlit
pandas
statsmodels
matplotlib
Pillow
